+++ src/routing/ors_wrapper.cpp
@@ -8,6 +8,7 @@ All rights reserved (see LICENSE).
 */
 
 #include "routing/ors_wrapper.h"
+#include "utils/helpers.h"
 
 namespace vroom::routing {
 
@@ -82,25 +83,16 @@ bool OrsWrapper::distance_value_is_null(
 
 UserDuration
 OrsWrapper::get_duration_value(const rapidjson::Value& matrix_entry) const {
-  return round_cost<UserDuration>(matrix_entry.GetDouble());
+  return utils::round<UserDuration>(matrix_entry.GetDouble());
 }
 
 UserDistance
 OrsWrapper::get_distance_value(const rapidjson::Value& matrix_entry) const {
-  return round_cost<UserDistance>(matrix_entry.GetDouble());
-}
-
-double OrsWrapper::get_total_distance(const rapidjson::Value& result) const {
-  return result["routes"][0]["summary"]["distance"].GetDouble();
+  return utils::round<UserDistance>(matrix_entry.GetDouble());
 }
 
 unsigned OrsWrapper::get_legs_number(const rapidjson::Value& result) const {
   return result["routes"][0]["segments"].Size();
-}
-
-double OrsWrapper::get_distance_for_leg(const rapidjson::Value& result,
-                                        rapidjson::SizeType i) const {
-  return result["routes"][0]["segments"][i]["distance"].GetDouble();
 }
 
 std::string OrsWrapper::get_geometry(rapidjson::Value& result) const {