+++ src/utils/output_json.h
@@ -12,33 +12,38 @@ All rights reserved (see LICENSE).
 
 #include "../include/rapidjson/include/rapidjson/document.h"
 #include "structures/vroom/solution/solution.h"
+#include "utils/exception.h"
 
 namespace vroom::io {
 
-rapidjson::Document to_json(const Solution& sol, bool geometry);
+rapidjson::Document to_json(const Solution& sol, bool report_distances);
+
+rapidjson::Document to_json(const vroom::Exception& e);
 
 rapidjson::Value to_json(const Summary& summary,
-                         bool geometry,
+                         bool report_distances,
                          rapidjson::Document::AllocatorType& allocator);
 
 rapidjson::Value to_json(const ComputingTimes& ct,
-                         bool geometry,
                          rapidjson::Document::AllocatorType& allocator);
 
 rapidjson::Value to_json(const Route& route,
-                         bool geometry,
+                         bool report_distances,
                          rapidjson::Document::AllocatorType& allocator);
 
 rapidjson::Value to_json(const Step& s,
-                         bool geometry,
+                         bool report_distances,
                          rapidjson::Document::AllocatorType& allocator);
 
 rapidjson::Value to_json(const Location& loc,
                          rapidjson::Document::AllocatorType& allocator);
 
+void write_to_json(const vroom::Exception& e,
+                   const std::string& output_file = "");
+
 void write_to_json(const Solution& sol,
                    const std::string& output_file = "",
-                   bool geometry = false);
+                   bool report_distances = false);
 
 } // namespace vroom::io