+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support for `max_distance` at vehicle level (#354)
+- `MAX_DISTANCE` violation cause in plan mode (#995)
 - Recommendation on how to cite in publications (#943)
 - Store distance matrices (#956)
 - Default radius of 35km for OSRM snapping (#922)
@@ -34,6 +35,7 @@
 - Refactor `RouteSplit` operator (#996)
 - Update cxxopts to 3.1.1 (#997)
 - Update gcc to version 12 in CI (#1002)
+- Distances in output are from internal matrices, not routing requests (#957)
 
 ### Fixed